requirements.txt
torch
Pillow
kagglehub
torchvision